org-platform/backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-spectacular==0.26.5
django-filter==24.2
requests==2.32.3
